refactor: 안쓰는 객체 삭제, 코드 포맷팅, repository로 패키지명 변경
