+++ src/main/java/kitchenpos/application/OrderService.java
@@ -5,10 +5,10 @@ import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.application.dtos.OrderRequest;
 import kitchenpos.application.dtos.OrderStatusRequest;
-import kitchenpos.dao.MenuRepository;
-import kitchenpos.dao.OrderLineItemRepository;
-import kitchenpos.dao.OrderRepository;
-import kitchenpos.dao.OrderTableRepository;
+import kitchenpos.repository.MenuRepository;
+import kitchenpos.repository.OrderLineItemRepository;
+import kitchenpos.repository.OrderRepository;
+import kitchenpos.repository.OrderTableRepository;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderLineItems;
