--- src/main/java/kitchenpos/dao/MenuDao.java
@@ -1,15 +1,0 @@
-package kitchenpos.dao;
-
-import java.util.List;
-import java.util.Optional;
-import kitchenpos.domain.Menu;
-
-public interface MenuDao {
-    Menu save(Menu entity);
-
-    Optional<Menu> findById(Long id);
-
-    List<Menu> findAll();
-
-    long countByIdIn(List<Long> ids);
-}
