--- src/main/java/kitchenpos/dao/OrderLineItemDao.java
@@ -1,15 +1,0 @@
-package kitchenpos.dao;
-
-import java.util.List;
-import java.util.Optional;
-import kitchenpos.domain.OrderLineItem;
-
-public interface OrderLineItemDao {
-    OrderLineItem save(OrderLineItem entity);
-
-    Optional<OrderLineItem> findById(Long id);
-
-    List<OrderLineItem> findAll();
-
-    List<OrderLineItem> findAllByOrderId(Long orderId);
-}
