--- src/main/java/kitchenpos/dao/OrderTableDao.java
@@ -1,17 +1,0 @@
-package kitchenpos.dao;
-
-import java.util.List;
-import java.util.Optional;
-import kitchenpos.domain.OrderTable;
-
-public interface OrderTableDao {
-    OrderTable save(OrderTable entity);
-
-    Optional<OrderTable> findById(Long id);
-
-    List<OrderTable> findAll();
-
-    List<OrderTable> findAllByIdIn(List<Long> ids);
-
-    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
-}
